requirements.txt
flask_sqlalchemy
flask
flask_cache
flask_validator
